add handler to reload existing wireguard interface
When new peers are added to the network, existing connections should not
be taken offline but interfaces should instead be edited in-flight.
This Commit adds and uses a handler that starts wireguard if not
running and afterwards reloads the peers from the configuration file
without taking down the wireguard interface.