Run hooks upon reconfiguration if necessary
WireGuards lifecycle hooks (Pre/PostUp/Down) are commonly used to manage
some additional system configuration. Without executing Down hooks prior
to and Up hooks after changes to them the system could be left in an un-
expected state.

For example, if we use WireGuard's lifecycle hooks to create and cleanup
some iptables rules we would miss to remove some of them if we don't run
the appropriate Down hooks before applying the configuration changes.
Likewise, the system won't end up in the expected state if WireGuard's
Up hooks aren't executed after updating them.

With this change WireGuard will be shutdown prior to reconfiguration if
any lifecycle hook will be affected. In fact, removals of Down hooks and
additions of Up hooks (both cover updates as well) will result in a re-
start of WireGuard which automatically executes the hooks in the correct
version.

This solves issue #135.